♻️ refactor: 엔티티 구조 변경
